Create password_cracker.py
I'm committing this now so you have the ability to use FindMatch() to test your code as you go. It basically accomplishes the dictionary attack method too. I'm not trying to steal @Janofermp 's part of the program, it just seemed like the simplest way to test my function. There is still probably some more functionality that can be added to the dictionary method.